requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
six==1.16.0
toml==0.10.2
